requirements.txt
Django==2.2.1
django-admin-sortable2==0.7.1
django-extensions==2.1.7
django-preferences==1.0.0
django-tinymce4-lite==1.7.5
jsmin==3.0.1
Pillow==9.3.0
psycopg2==2.9.5
pytz==2019.1
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
